+++ backend/platform/src/main/java/ru/abondin/hreasy/platform/service/admin/employee/dto/CurrentProjectTransferApproverDto.java
@@ -1,0 +1,11 @@
+package ru.abondin.hreasy.platform.service.admin.employee.dto;
+
+import lombok.Data;
+
+@Data
+public class CurrentProjectTransferApproverDto {
+    private Integer employeeId;
+    private String displayName;
+    private String email;
+    private String managerType;
+}